@@ -11,11 +11,10 @@ description-file = "README.rst"
 classifiers = [
 "License :: OSI Approved :: MIT License",
 "Programming Language :: Python :: 3",
-"Programming Language :: Python :: 3.5",
 "Programming Language :: Python :: 3.6",
 "Programming Language :: Python :: 3.7"
 ]
-requires-python = ">=3.5"
+requires-python = ">=3.6"
 requires = [
 "click",
 "eventlet",
