@@ -9,13 +9,12 @@ def test_markdown():
     next_uuid = Markdown._NEXT_UUID
     mark = Markdown()
     assert mark._instantiate == (
-        "<Markdown initial={{''}} "
-        "socket={{socket}} uuid={{'{next_uuid}'}} />"
-    ).format(next_uuid=next_uuid + 1)
+        f"<Markdown initial={{''}} socket={{socket}} uuid={{'{next_uuid + 1}'}} />"
+    )
 
     next_uuid = Markdown._NEXT_UUID
     mark = Markdown(initial='#hi\n##hello')
     assert mark._instantiate == (
-        "<Markdown initial={{'<h1>hi</h1>\\n<h2>hello</h2>'}} "
-        "socket={{socket}} uuid={{'{next_uuid}'}} />"
-    ).format(next_uuid=next_uuid + 1)
+        "<Markdown initial={'<h1>hi</h1>\\n<h2>hello</h2>'} "
+        f"socket={{socket}} uuid={{'{next_uuid + 1}'}} />"
+    )
